+++ src/main/java/visang/showcase/aibackend/service/StudyService.java
@@ -103,14 +103,14 @@ public class StudyService {
 
 
         // 세션에 저장된 진단평가 100문항에서 필요한 정보 사용
-        List<Integer> q_idx_list = probList.stream()
-                .map(m -> m.getQ_idx()).collect(Collectors.toList());  // 토픽 리스트
+        List<Long> q_idx_list = probList.stream()
+                .map(m -> Long.valueOf(m.getQ_idx())).collect(Collectors.toList());  // 토픽 리스트
 
-        List<Integer> correct_list = probList.stream()
-                .map(m -> m.getCorrect()).collect(Collectors.toList()); // 정오답 리스트
+        List<Long> correct_list = probList.stream()
+                .map(m -> Long.valueOf(m.getCorrect())).collect(Collectors.toList()); // 정오답 리스트
 
-        List<Integer> diff_level_list = probList.stream()
-                .map(m -> m.getDiff_level()).collect(Collectors.toList()); // 문제 난이도 리스트
+        List<Long> diff_level_list = probList.stream()
+                .map(m -> Long.valueOf(m.getDiff_level())).collect(Collectors.toList()); // 문제 난이도 리스트
 
         // INPUT__ 객체 생성
         List<KnowledgeReqObject> inputs = new ArrayList<>();
@@ -119,7 +119,7 @@ public class StudyService {
         inputs.add(createRequestObj(2, diff_level_list));
 
         // 옵션 값 추가
-        inputs.add(createRequestObj(3, List.of(tgtTopic, 5, 1, 5, 5)));
+        inputs.add(createRequestObj(3, List.of(Long.valueOf(tgtTopic), 5L, 1L, 5L, 5L)));
 
         return new KnowledgeLevelRequest(inputs);
     }
@@ -127,7 +127,7 @@ public class StudyService {
     /**
      * RequestBody의 INPUT__ 요청 객체 생성
      */
-    private KnowledgeReqObject createRequestObj(int idx, List<Integer> payload) {
+    private KnowledgeReqObject createRequestObj(int idx, List<Long> payload) {
         KnowledgeReqObject obj = new KnowledgeReqObject();
         obj.setName("INPUT__" + idx);
         obj.setDatatype("INT64");
@@ -135,7 +135,7 @@ public class StudyService {
         List<Integer> shape = List.of(1, payload.size()); // 배치 사이즈, 문항 수
         obj.setShape(shape);
 
-        List<List<Integer>> data = List.of(payload);
+        List<List<Long>> data = List.of(payload);
         obj.setData(data);
 
         return obj;
